api/requirements.txt
# Fichier : requirements.txt (à la racine du projet)

# Serveur ASGI et gestionnaire de processus de production
gunicorn
uvicorn[standard]

# Framework pour l'API
fastapi
python-multipart
selenium

# Bibliothèques d'analyse financière et de données
yfinance
pandas
numpy
lxml

# Pydantic est une dépendance de FastAPI, mais il est bon de l'expliciter
pydantic

# Base de données et cache
redis
python-dotenv

# Gestion des variables d'environnement
python-dotenv

requests

# Bibliothèque pour le scraping de données financières
stockdex